Restructure packages, creating a new processor package
