+++ reasoner/common/Tracer.java
@@ -20,8 +20,8 @@ package com.vaticle.typedb.core.reasoner.common;
 
 import com.vaticle.typedb.core.common.exception.TypeDBException;
 import com.vaticle.typedb.core.concurrent.actor.Actor;
-import com.vaticle.typedb.core.reasoner.reactive.Monitor;
-import com.vaticle.typedb.core.reasoner.reactive.Reactive.Identifier;
+import com.vaticle.typedb.core.reasoner.processor.reactive.Monitor;
+import com.vaticle.typedb.core.reasoner.processor.reactive.Reactive.Identifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
