+++ reasoner/controller/ConjunctionController.java
@@ -30,13 +30,14 @@ import com.vaticle.typedb.core.pattern.Conjunction;
 import com.vaticle.typedb.core.reasoner.answer.Mapping;
 import com.vaticle.typedb.core.reasoner.common.Planner;
 import com.vaticle.typedb.core.reasoner.controller.ControllerRegistry.ControllerView;
-import com.vaticle.typedb.core.reasoner.reactive.AbstractProcessor;
-import com.vaticle.typedb.core.reasoner.reactive.AbstractProcessor.Connector;
-import com.vaticle.typedb.core.reasoner.reactive.AbstractProcessor.Connector.AbstractRequest;
-import com.vaticle.typedb.core.reasoner.reactive.Reactive;
-import com.vaticle.typedb.core.reasoner.reactive.Reactive.Publisher;
-import com.vaticle.typedb.core.reasoner.reactive.TransformationStream;
-import com.vaticle.typedb.core.reasoner.reactive.common.Operator;
+import com.vaticle.typedb.core.reasoner.processor.AbstractProcessor;
+import com.vaticle.typedb.core.reasoner.processor.Connector;
+import com.vaticle.typedb.core.reasoner.processor.Connector.AbstractRequest;
+import com.vaticle.typedb.core.reasoner.processor.Input;
+import com.vaticle.typedb.core.reasoner.processor.reactive.Reactive;
+import com.vaticle.typedb.core.reasoner.processor.reactive.Reactive.Publisher;
+import com.vaticle.typedb.core.reasoner.processor.reactive.TransformationStream;
+import com.vaticle.typedb.core.reasoner.processor.reactive.common.Operator;
 import com.vaticle.typedb.core.traversal.common.Identifier.Variable;
 
 import java.util.ArrayList;
