+++ reasoner/processor/reactive/common/SubscriberRegistry.java
@@ -16,9 +16,9 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.reactive.common;
+package com.vaticle.typedb.core.reasoner.processor.reactive.common;
 
-import com.vaticle.typedb.core.reasoner.reactive.Reactive.Subscriber;
+import com.vaticle.typedb.core.reasoner.processor.reactive.Reactive.Subscriber;
 
 import java.util.HashSet;
 import java.util.Set;
