--- reasoner/reactive/AbstractProcessor.java
@@ -1,443 +1,0 @@
-/*
- * Copyright (C) 2021 Vaticle
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package com.vaticle.typedb.core.reasoner.reactive;
-
-import com.vaticle.typedb.common.collection.Pair;
-import com.vaticle.typedb.core.common.exception.TypeDBException;
-import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
-import com.vaticle.typedb.core.concurrent.actor.Actor;
-import com.vaticle.typedb.core.reasoner.common.Tracer;
-import com.vaticle.typedb.core.reasoner.controller.AbstractController;
-import com.vaticle.typedb.core.reasoner.reactive.AbstractProcessor.Connector.AbstractRequest;
-import com.vaticle.typedb.core.reasoner.reactive.Reactive.Identifier;
-import com.vaticle.typedb.core.reasoner.reactive.Reactive.Publisher;
-import com.vaticle.typedb.core.reasoner.reactive.Reactive.Subscriber;
-import com.vaticle.typedb.core.reasoner.reactive.common.PublisherDelegate;
-import com.vaticle.typedb.core.reasoner.reactive.common.ReactiveIdentifier;
-import com.vaticle.typedb.core.reasoner.reactive.common.SubscriberDelegate;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.Supplier;
-
-import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_OPERATION;
-import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
-import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.RESOURCE_CLOSED;
-
-public abstract class AbstractProcessor<INPUT, OUTPUT,
-        REQ extends AbstractRequest<?, ?, INPUT>,
-        REACTIVE_BLOCK extends AbstractProcessor<INPUT, OUTPUT, REQ, REACTIVE_BLOCK>> extends Actor<REACTIVE_BLOCK> {
-
-    private static final Logger LOG = LoggerFactory.getLogger(AbstractProcessor.class);
-
-    private final Driver<? extends AbstractController<?, INPUT, OUTPUT, REQ, REACTIVE_BLOCK, ?>> controller;
-    private final Context context;
-    private final Map<Identifier<?, ?>, Input<INPUT>> inputs;  // TODO: inputPorts (sweeping rename)
-    private final Map<Identifier<?, ?>, Output<OUTPUT>> outputs;
-    private final Map<Pair<Identifier<?, ?>, Identifier<?, ?>>, Runnable> pullRetries;
-    private Reactive.Stream<OUTPUT,OUTPUT> initialReactive;
-    private boolean terminated;
-    private long reactiveCounter;
-
-    protected AbstractProcessor(Driver<REACTIVE_BLOCK> driver,
-                                Driver<? extends AbstractController<?, INPUT, OUTPUT, REQ, REACTIVE_BLOCK, ?>> controller,
-                                Context context, Supplier<String> debugName) {
-        super(driver, debugName);
-        this.controller = controller;
-        this.context = context;
-        this.inputs = new HashMap<>();
-        this.outputs = new HashMap<>();
-        this.reactiveCounter = 0;
-        this.pullRetries = new HashMap<>();
-    }
-
-    public abstract void setUp();
-
-    protected void setInitialReactive(Reactive.Stream<OUTPUT, OUTPUT> initialReactive) {
-        this.initialReactive = initialReactive;
-    }
-
-    public Reactive.Stream<OUTPUT,OUTPUT> outputRouter() {  // TODO: initialReactive or hubReactive
-        return initialReactive;
-    }
-
-    public void rootPull() {
-        throw TypeDBException.of(ILLEGAL_OPERATION);
-    }
-
-    public void pull(Identifier<?, ?> outputId) {
-        outputs.get(outputId).pull();
-    }
-
-    public void receive(Identifier<?, ?> inputId, INPUT packet, Identifier<?, INPUT> publisherId) {
-        inputs.get(inputId).receive(publisherId, packet);
-    }
-
-    public <PACKET> void schedulePullRetry(Publisher<PACKET> publisher, Subscriber<PACKET> subscriber) {
-        pullRetries.put(new Pair<>(publisher.identifier(), subscriber.identifier()), () -> publisher.pull(subscriber));
-        driver().execute(actor -> actor.pullRetry(publisher.identifier(), subscriber.identifier()));
-    }
-
-    protected void pullRetry(Identifier<?, ?> publisher, Identifier<?, ?> subscriber) {
-        tracer().ifPresent(tracer -> tracer.pullRetry(subscriber, publisher));
-        pullRetries.get(new Pair<Identifier<?, ?>, Identifier<?, ?>>(publisher, subscriber)).run();
-    }
-
-    protected void requestConnection(REQ req) {
-        if (isTerminated()) return;
-        controller.execute(actor -> actor.routeConnectionRequest(req));
-    }
-
-    public void establishConnection(Connector<?, OUTPUT> connector) {
-        if (isTerminated()) return;
-        Output<OUTPUT> output = createOutput();
-        output.setSubscriber(connector.inputId());
-        connector.connectViaTransforms(outputRouter(), output);
-        connector.inputId().processor().execute(
-                actor -> actor.finishConnection(connector.inputId(), output.identifier()));
-    }
-
-    protected void finishConnection(Identifier<INPUT, ?> inputId, Identifier<?, INPUT> outputId) {
-        Input<INPUT> input = inputs.get(inputId);
-        input.setOutput(outputId);
-        input.pull();
-    }
-
-    protected Input<INPUT> createInput() {
-        Input<INPUT> input = new Input<>(this);
-        inputs.put(input.identifier(), input);
-        return input;
-    }
-
-    protected Output<OUTPUT> createOutput() {
-        Output<OUTPUT> output = new Output<>(this);
-        outputs.put(output.identifier(), output);
-        return output;
-    }
-
-    public Driver<Monitor> monitor() {
-        return context.monitor();
-    }
-
-    Optional<Tracer> tracer() {
-        return context.tracer();
-    }
-
-    public Context context() {
-        return context;
-    }
-
-    public void onFinished(Identifier<?, ?> finishable) {
-        throw TypeDBException.of(ILLEGAL_STATE);
-    }
-
-    @Override
-    protected void exception(Throwable e) {
-        if (e instanceof TypeDBException && ((TypeDBException) e).code().isPresent()) {
-            String code = ((TypeDBException) e).code().get();
-            if (code.equals(RESOURCE_CLOSED.code())) {
-                LOG.debug("Processor interrupted by resource close: {}", e.getMessage());
-                controller.execute(actor -> actor.exception(e));
-                return;
-            } else {
-                LOG.debug("Processor interrupted by TypeDB exception: {}", e.getMessage());
-            }
-        }
-        LOG.error("Actor exception", e);
-        controller.execute(actor -> actor.exception(e));
-    }
-
-    public void terminate(Throwable cause) {
-        LOG.debug("Actor terminated.", cause);
-        this.terminated = true;
-    }
-
-    public boolean isTerminated() {
-        return terminated;
-    }
-
-    public long incrementReactiveCounter() {
-        reactiveCounter += 1;
-        return reactiveCounter;
-    }
-
-    public Identifier<INPUT, OUTPUT> registerReactive(Reactive reactive) {
-        return new ReactiveIdentifier<>(driver(), reactive, incrementReactiveCounter());
-    }
-
-    public static class Connector<BOUNDS, PACKET> {
-
-        private final Identifier<PACKET, ?> inputId;
-        private final List<Function<PACKET, PACKET>> transforms;
-        private final BOUNDS bounds;
-
-        public Connector(Identifier<PACKET, ?> inputId, BOUNDS bounds) {
-            this.inputId = inputId;
-            this.transforms = new ArrayList<>();
-            this.bounds = bounds;
-        }
-
-        public Connector(Identifier<PACKET, ?> inputId, BOUNDS bounds, List<Function<PACKET, PACKET>> transforms) {
-            this.inputId = inputId;
-            this.transforms = transforms;
-            this.bounds = bounds;
-        }
-
-        public void connectViaTransforms(Reactive.Stream<PACKET, PACKET> toConnect, Output<PACKET> output) {
-            Publisher<PACKET> op = toConnect;
-            for (Function<PACKET, PACKET> t : transforms) op = op.map(t);
-            op.registerSubscriber(output);
-        }
-
-        public BOUNDS bounds(){
-            return bounds;
-        }
-
-        public Identifier<PACKET, ?> inputId() {
-            return inputId;
-        }
-
-        public Connector<BOUNDS, PACKET> withMap(Function<PACKET, PACKET> function) {
-            ArrayList<Function<PACKET, PACKET>> newTransforms = new ArrayList<>(transforms);
-            newTransforms.add(function);
-            return new Connector<>(inputId, bounds, newTransforms);
-        }
-
-        public Connector<BOUNDS, PACKET> withNewBounds(BOUNDS newBounds) {
-            return new Connector<>(inputId, newBounds, transforms);
-        }
-
-        public abstract static class AbstractRequest<CONTROLLER_ID, BOUNDS, PACKET> {
-
-            // TODO: Should hold on to the processor that sent the request
-            private final Identifier<PACKET, ?> inputId;  // TODO: requesterPortID
-            private final CONTROLLER_ID controllerId;
-            private final BOUNDS bounds;
-
-            protected AbstractRequest(Identifier<PACKET, ?> inputId, CONTROLLER_ID controllerId,
-                                      BOUNDS bounds) {
-                this.inputId = inputId;
-                this.controllerId = controllerId;
-                this.bounds = bounds;
-            }
-
-            public Identifier<PACKET, ?> inputId() {
-                return inputId;
-            }
-
-            public CONTROLLER_ID controllerId() {
-                return controllerId;
-            }
-
-            public BOUNDS bounds() {
-                return bounds;
-            }
-
-            @Override
-            public boolean equals(Object o) {
-                // TODO: be wary with request equality when conjunctions are involved
-                if (this == o) return true;
-                if (o == null || getClass() != o.getClass()) return false;
-                AbstractRequest<?, ?, ?> request = (AbstractRequest<?, ?, ?>) o;
-                return inputId == request.inputId &&
-                        controllerId.equals(request.controllerId) &&
-                        bounds.equals(request.bounds);
-            }
-
-            @Override
-            public int hashCode() {
-                return Objects.hash(controllerId, inputId, bounds);
-            }
-
-        }
-    }
-
-    public static class Context {
-
-        private final Driver<Monitor> monitor;
-        private final Tracer tracer;
-
-        public Context(Driver<Monitor> monitor, @Nullable Tracer tracer) {
-            this.monitor = monitor;
-            this.tracer = tracer;
-        }
-
-        public Optional<Tracer> tracer() {
-            return Optional.ofNullable(tracer);
-        }
-
-        public Driver<Monitor> monitor() {
-            return monitor;
-        }
-
-    }
-
-    /**
-     * Governs an input to a processor
-     */
-    public static class Input<PACKET> implements Publisher<PACKET> {
-
-        private final Identifier<PACKET, ?> identifier;
-        private final AbstractProcessor<PACKET, ?, ?, ?> processor;
-        private final PublisherDelegate<PACKET> publisherActions;
-        private boolean ready;
-        private Identifier<?, PACKET> providingOutput;  // TODO: Output PortID
-        private Subscriber<PACKET> subscriber;
-
-        public Input(AbstractProcessor<PACKET, ?, ?, ?> processor) {
-            this.processor = processor;
-            this.identifier = processor.registerReactive(this);
-            this.ready = false;
-            this.publisherActions = new PublisherDelegate<>(this, processor.context());
-        }
-
-        @Override
-        public AbstractProcessor<?, ?, ?, ?> processor() {
-            return processor;
-        }
-
-        @Override
-        public Identifier<PACKET, ?> identifier() {
-            return identifier;
-        }
-
-        public void setOutput(Identifier<?, PACKET> outputId) {
-            assert providingOutput == null;
-            providingOutput = outputId;
-            processor().monitor().execute(actor -> actor.registerPath(identifier(), outputId));
-            assert !ready;
-            ready = true;
-        }
-
-        public void pull() {
-            pull(subscriber);
-        }
-
-        @Override
-        public void pull(Subscriber<PACKET> subscriber) {
-            assert subscriber.equals(this.subscriber);
-            processor().tracer().ifPresent(tracer -> tracer.pull(subscriber.identifier(), identifier()));
-            if (ready) providingOutput.processor().execute(actor -> actor.pull(providingOutput));  // TODO: Store the processor rather than getting it from the outputPort
-        }
-
-        @Override
-        public void registerSubscriber(Subscriber<PACKET> subscriber) {
-            assert this.subscriber == null;
-            this.subscriber = subscriber;
-            subscriber.registerPublisher(this);
-        }
-
-        @Override
-        public <MAPPED> Stream<PACKET, MAPPED> map(Function<PACKET, MAPPED> function) {
-            return publisherActions.map(this, function);
-        }
-
-        @Override
-        public <MAPPED> Stream<PACKET, MAPPED> flatMap(Function<PACKET, FunctionalIterator<MAPPED>> function) {
-            return publisherActions.flatMap(this, function);
-        }
-
-        @Override
-        public Stream<PACKET, PACKET> buffer() {
-            return publisherActions.buffer(this);
-        }
-
-        @Override
-        public Stream<PACKET, PACKET> distinct() {
-            return publisherActions.distinct(this);
-        }
-
-        public void receive(Identifier<?, PACKET> outputId, PACKET packet) {
-            processor().tracer().ifPresent(tracer -> tracer.receive(outputId, identifier(), packet));
-            subscriber.receive(this, packet);
-        }
-
-        @Override
-        public String toString() {
-            return processor.debugName().get() + ":" + getClass().getSimpleName();
-        }
-
-    }
-
-    /**
-     * Governs an output from a processor
-     */
-    public static class Output<PACKET> implements Subscriber<PACKET> {
-
-        private final Identifier<?, PACKET> identifier;
-        private final AbstractProcessor<?, PACKET, ?, ?> processor;
-        private final SubscriberDelegate<PACKET> subscriberActions;
-        private Identifier<PACKET, ?> receivingInput;
-        private Publisher<PACKET> publisher;
-
-        public Output(AbstractProcessor<?, PACKET, ?, ?> processor) {
-            this.processor = processor;
-            this.identifier = processor().registerReactive(this);
-            this.subscriberActions = new SubscriberDelegate<>(this, processor().context());
-        }
-
-        @Override
-        public Identifier<?, PACKET> identifier() {
-            return identifier;
-        }
-
-        @Override
-        public AbstractProcessor<?, PACKET, ?, ?> processor() {
-            return processor;
-        }
-
-        @Override
-        public void receive(Publisher<PACKET> publisher, PACKET packet) {
-            subscriberActions.traceReceive(publisher, packet);
-            receivingInput.processor().execute(actor -> actor.receive(receivingInput, packet, identifier()));
-        }
-
-        public void pull() {
-            assert publisher != null;
-            processor().context().tracer().ifPresent(tracer -> tracer.pull(receivingInput, identifier()));
-            publisher.pull(this);
-        }
-
-        @Override
-        public void registerPublisher(Publisher<PACKET> publisher) {
-            assert this.publisher == null;
-            this.publisher = publisher;
-            subscriberActions.registerPath(publisher);
-        }
-
-        public void setSubscriber(Identifier<PACKET, ?> inputId) {
-            assert receivingInput == null;
-            receivingInput = inputId;
-        }
-
-        @Override
-        public String toString() {
-            return processor().debugName().get() + ":" + getClass().getSimpleName();
-        }
-    }
-
-}
